+++ pkg/common/record.go
@@ -1,0 +1,9 @@
+package common
+
+type Record struct {
+	Type    string
+	Name    string
+	Target  string
+	TTL     int
+	Proxied bool
+}